Refactor: change name of subscribers
